requirements.txt
flask==3.0.0
anthropic==0.40.0
arxiv==2.1.3
apscheduler==3.10.4
sqlalchemy==2.0.25
python-dotenv==1.0.0
